refactor: split balance & metrics god modules -> grade A (honest, codex)
Behavior-preserving split of balance.rs and metrics.rs into cohesive submodule
directories (src/balance/{mod,score,issues,grade,rationale,patterns,...}.rs and
src/metrics/), public API stable via re-exports. Internal callers depend on
focused submodules instead of the broad facade.

Effect: God Module + high-afferent hubs cleared; cascading to balance resolved;
'coupling ./src' now grades A (0 high/critical, score 0.76) with full git
analysis and an unmodified .coupling.toml — reached by genuine architecture,
not gaming.
