+++ src/balance/issues.rs
@@ -1,0 +1,7 @@
+//! Compatibility re-exports for the balance issue model.
+
+pub use super::action::RefactoringAction;
+pub use super::issue::CouplingIssue;
+pub use super::issue_type::IssueType;
+pub use super::rationale::{GradeDimension, GradeRationale, IssueTypeContribution};
+pub use super::severity::Severity;
