+++ src/balance/severity.rs
@@ -1,0 +1,23 @@
+/// Issue severity levels
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub enum Severity {
+    /// Minor issue, consider addressing
+    Low,
+    /// Should be addressed in regular maintenance
+    Medium,
+    /// Needs attention soon, potential source of bugs
+    High,
+    /// Must be fixed, actively causing problems
+    Critical,
+}
+
+impl std::fmt::Display for Severity {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Severity::Low => write!(f, "Low"),
+            Severity::Medium => write!(f, "Medium"),
+            Severity::High => write!(f, "High"),
+            Severity::Critical => write!(f, "Critical"),
+        }
+    }
+}
